# Optimize startup time