--- a/Common/src/main/java/com/almostreliable/unified/config/DuplicateConfig.java
+++ b/Common/src/main/java/com/almostreliable/unified/config/DuplicateConfig.java
@@ -60,18 +60,33 @@ public final class DuplicateConfig extends Config {
      * @return True if the recipe type is ignored, false otherwise
      */
     private boolean isRecipeTypeIgnored(RecipeLink recipe) {
-        return ignoredRecipeTypesCache.computeIfAbsent(recipe.getType(), type -> {
-            for (Pattern ignorePattern : ignoreRecipeTypes) {
-                if (ignorePattern.matcher(type.toString()).matches()) {
-                    return true;
-                }
+        ResourceLocation type = recipe.getType();
+        Boolean ignored = ignoredRecipeTypesCache.get(type);
+        if (ignored == null) {
+            ignored = computeIsRecipeTypeIgnored(recipe);
+            ignoredRecipeTypesCache.put(type, ignored);
+        }
+        return ignored;
+    }
+
+    private boolean computeIsRecipeTypeIgnored(RecipeLink recipe) {
+        ResourceLocation type = recipe.getType();
+        String typeString = type.toString();
+        for (Pattern ignorePattern : ignoreRecipeTypes) {
+            if (ignorePattern.matcher(typeString).matches()) {
+                return true;
             }
-            return false;
-        });
+        }
+        return false;
     }
 
     public JsonCompare.CompareSettings getCompareSettings(ResourceLocation type) {
         return overrideRules.getOrDefault(type, defaultRules);
+    }
+
+    public JsonCompare.CompareContext getCompareContext(RecipeLink recipe) {
+        JsonCompare.CompareSettings compareSettings = getCompareSettings(recipe.getType());
+        return JsonCompare.CompareContext.create(compareSettings, recipe);
     }
 
     public boolean shouldCompareAll() {