--- a/Common/src/main/java/com/almostreliable/unified/unification/recipe/RecipeLink.java
+++ b/Common/src/main/java/com/almostreliable/unified/unification/recipe/RecipeLink.java
@@ -11,11 +11,19 @@ import com.google.gson.JsonObject;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 
 public class RecipeLink implements RecipeData {
+    /**
+     * This cache is an optimization to avoid creating many ResourceLocations for just a few different types.
+     * Having fewer ResourceLocation instances can greatly speed up equality checking when these are used as map keys.
+     */
+    private static final Map<String, ResourceLocation> PARSED_TYPE_CACHE = new HashMap<>();
+
     private final ResourceLocation id;
     private final ResourceLocation type;
     private final JsonObject originalRecipe;
@@ -27,7 +35,8 @@ public class RecipeLink implements RecipeData {
         this.originalRecipe = originalRecipe;
 
         try {
-            this.type = ResourceLocation.tryParse(originalRecipe.get("type").getAsString());
+            String typeString = originalRecipe.get("type").getAsString();
+            this.type = PARSED_TYPE_CACHE.computeIfAbsent(typeString, ResourceLocation::parse);
         } catch (Exception e) {
             throw new IllegalArgumentException("could not detect recipe type");
         }
@@ -40,19 +49,19 @@ public class RecipeLink implements RecipeData {
      *
      * @param first           first recipe to compare
      * @param second          second recipe to compare
-     * @param compareSettings Settings to use for comparison.
+     * @param compareContext  Settings and context to use for comparison.
      * @return the recipe where rules are applied and the recipes are compared for equality, or null if the recipes are not equal
      */
     @Nullable
-    public static RecipeLink compare(RecipeLink first, RecipeLink second, JsonCompare.CompareSettings compareSettings) {
+    public static RecipeLink compare(RecipeLink first, RecipeLink second, JsonCompare.CompareContext compareContext) {
         JsonObject selfActual = first.getActual();
         JsonObject toCompareActual = second.getActual();
 
         JsonObject compare = null;
         if (first.getType().toString().equals("minecraft:crafting_shaped")) {
-            compare = JsonCompare.compareShaped(selfActual, toCompareActual, compareSettings);
-        } else if (JsonCompare.matches(selfActual, toCompareActual, compareSettings)) {
-            compare = JsonCompare.compare(compareSettings.getRules(), selfActual, toCompareActual);
+            compare = JsonCompare.compareShaped(selfActual, toCompareActual, compareContext);
+        } else if (JsonCompare.matches(selfActual, toCompareActual, compareContext)) {
+            compare = JsonCompare.compare(compareContext.settings().getRules(), selfActual, toCompareActual);
         }
 
         if (compare == null) return null;
@@ -129,10 +138,10 @@ public class RecipeLink implements RecipeData {
      * the master from the link will be used. Otherwise, we will create a new link if needed.
      *
      * @param otherRecipe     Recipe data to check for duplicate against.
-     * @param compareSettings Settings to use for comparison.
+     * @param compareContext  Settings and context to use for comparison.
      * @return True if recipe is a duplicate, false otherwise.
      */
-    public boolean handleDuplicate(RecipeLink otherRecipe, JsonCompare.CompareSettings compareSettings) {
+    public boolean handleDuplicate(RecipeLink otherRecipe, JsonCompare.CompareContext compareContext) {
         DuplicateLink selfDuplicate = getDuplicateLink();
         DuplicateLink otherDuplicate = otherRecipe.getDuplicateLink();
 
@@ -141,7 +150,7 @@ public class RecipeLink implements RecipeData {
         }
 
         if (selfDuplicate == null && otherDuplicate == null) {
-            RecipeLink compare = compare(this, otherRecipe, compareSettings);
+            RecipeLink compare = compare(this, otherRecipe, compareContext);
             if (compare == null) {
                 return false;
             }
@@ -153,7 +162,7 @@ public class RecipeLink implements RecipeData {
         }
 
         if (otherDuplicate != null) {
-            RecipeLink compare = compare(this, otherDuplicate.getMaster(), compareSettings);
+            RecipeLink compare = compare(this, otherDuplicate.getMaster(), compareContext);
             if (compare == null) {
                 return false;
             }
@@ -163,7 +172,7 @@ public class RecipeLink implements RecipeData {
         }
 
         // selfDuplicate != null
-        RecipeLink compare = compare(selfDuplicate.getMaster(), otherRecipe, compareSettings);
+        RecipeLink compare = compare(selfDuplicate.getMaster(), otherRecipe, compareContext);
         if (compare == null) {
             return false;
         }