--- a/Common/src/main/java/com/almostreliable/unified/unification/recipe/RecipeTransformer.java
+++ b/Common/src/main/java/com/almostreliable/unified/unification/recipe/RecipeTransformer.java
@@ -138,7 +138,8 @@ public class RecipeTransformer {
             return false;
         }
 
-        JsonCompare.CompareSettings compareSettings = duplicateConfig.getCompareSettings(curRecipe.getType());
+        JsonCompare.CompareContext compareContext = duplicateConfig.getCompareContext(curRecipe);
+
         boolean foundDuplicate = false;
         for (RecipeLink recipeLink : recipes) {
             if (!curRecipe.getType().equals(recipeLink.getType())) {
@@ -150,7 +151,7 @@ public class RecipeTransformer {
                 continue;
             }
 
-            foundDuplicate |= curRecipe.handleDuplicate(recipeLink, compareSettings);
+            foundDuplicate |= curRecipe.handleDuplicate(recipeLink, compareContext);
         }
 
         return foundDuplicate;