Hello!

## Proposed Changes

This PR optimizes some very hot areas when starting AlmostUnified in a large pack.
I tested this against ATM10 and was able to speed up this mod's loading by about 40%, from 9.5 seconds to 5.7 seconds on my machine. This is a really imprecise measurement but I am confident that it will be at least a little faster for everyone.

## Additional Context

### Future Work

Almost all the remaining time is spent in `com.google.gson.JsonObject#equals`, but trying to avoid that would require a big structural change so I stopped here. I think using java maps of values instead of `JsonObject` for `RecipeLink#getActual` could give a massive speedup, but there are probably things here I haven't considered so I did not try it.

### Performance Comparison (flame graphs)

Before:

(9.5 seconds)
![Screenshot 2024-10-09 at 9 03 55 AM](https://github.com/user-attachments/assets/b95901e2-8d43-4616-ab92-e596a1f3caf6)

After:

(5.7 seconds)
![Screenshot 2024-10-09 at 9 03 40 AM](https://github.com/user-attachments/assets/e3a7521b-7be5-4995-889b-cd6210a8cb04)