@rlnt 2024-10-16
Thanks a lot for this pull request and the explanation. Since I don't have a lot of time right now, it might take me a bit to look at this more closely. A lot of this is new to me. I would have assumed the JIT optimizes things like the lambda. And I also need to learn how to read flame graphs. 😄

About the equality checks of `JsonObject`s. I would need some more context on your suggestion of using Java maps. As far as I understand it, `JsonObject`s are also just map-like objects and this shouldn't make a huge difference.
We were also not happy with the approach of comparing the JSON structures just to tell if something had changed. But when working on the 1.0 update for Minecraft 1.21 and above, we implemented a new approach with something we called a `JsonCursor`. That approach used partial comparisons only and we even tried another approach of methods telling us whether something in the JSON changed to avoid comparison at all. However, this made no noticeable performance difference and we reverted to comparison.

The comparison also has the huge advantage of simplicity. With a proper API introduced in 1.0, mod authors can register their own custom unifiers that tell Almost Unified how to treat their recipe JSONs. Since we compare the JSON afterward, we can tell whether something changed without it being a burden for mod authors to track whether their unifier changed something.