@mezz 2024-10-18
No rush!

A flame graph is like a stack trace. The bottom is the first parent method, and then on top of it are the children methods called from it.
The children methods are sorted left to right so that the ones that take more time are on the left.

The capturing lambda is pretty fast in Java but it has to create something like an object in order to store (capture) the current state outside of itself when it is created, in order to access the variables it wants to use when it is called. Normally you can ignore the performance hit, it really doesn't matter, but in this specific case it's being called an incredible number of times so it actually has some measurable impact.

`JsonObject` is great here because it does exactly what you want to do and does it simply, I think you had the right idea to use it. Replacing it with anything better may be more work. 
Unfortunately Gson does something a bit lazy with the values by design, they're stored in a way that's really generic so `equals` and `hashcode` comparisons have a lot of extra overhead. Each value is put into a `JsonPrimitive` which could be a `String`, or any primitive like `Integer`, `Long`, `Double`, etc.
Take a look at the `equals` method and you'll see why it can be a bit slow when called a bajillion times in a row:
https://github.com/google/gson/blob/2ce6a6106c22bf0cce2e39f68b0da2f117d6e66b/gson/src/main/java/com/google/gson/JsonPrimitive.java#L267-L295
It's sort of parsing the values every time you want to do something with the `JsonElement`.

We can do better than the Gson implementation because we already know the actual type of the values after parsing the Json, so we can create and do comparisons on a simple `Map<String, Object>` (where the `Object` is parsed value like an `Integer` or `String`), which is more efficient than doing comparisons on `JsonObject`'s internal `Map<String, JsonElement>`.